rename getBubbleSizFn -> makeBubbleSizeFn + put it separate file
